confidence -- --recent option
This option provides a hacky approach for dealing with
changing data.

Get some idea of what the current population mean is
from a timeseries, by searching back until the
confidence interval is quite small.

The lower the tolerance and greater your confidence is
the older your mean will be (in general), but the
more certainty you will have.

There are some multicomparison considerations here
which I ignore.

For alternative or more rigorous approaches you
might be interested in Kalman filters or
changepoint detection (see the changepoint) library
in R.